# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OGM Gönüllü Yönetim Sistemi API",
    "version": "0.1.0",
    "description": "T.C. Orman Genel Müdürlüğü Gönüllü Yönetim Sistemi backend API'si. Auth (e-Devlet + Telefon OTP), onboarding, profil, ana sayfa, yangın bildirimi ve referans veri uçlarını içerir.",
    "contact": {
      "name": "OGM Gönüllü Backend",
      "email": "gonullu-sistem@ogm.gov.tr"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://surfaces-acer-munich-completed.trycloudflare.com/v1",
      "description": "API v1"
    },
    {
      "url": "https://surfaces-acer-munich-completed.trycloudflare.com",
      "description": "Kök (health vb.)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Erişim token'ı (`Authorization: Bearer <token>`)"
      },
      "registrationToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Telefon doğrulama sonrası kayıt formu için kısa süreli token"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Panel admin anahtarı (ADMIN_API_KEY)"
      },
      "officerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Saha amiri anahtarı (OFFICER_API_KEY)"
      }
    },
    "schemas": {
...
      "CompletedTraining": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "durationMin": {
            "type": "integer"
          },
          "completedAt": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "instructorName": {
            "type": "string",
            "nullable": true
          },
          "progressPercent": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "in_progress"
            ]
          },
          "certificateUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          }
        }
      },
...
    },
...
  },
  "security": [],
  "paths": {
...
    "/users/me/trainings": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "summary": "Aldığım eğitimler",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tamamlanan/sürdürülen eğitimler",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompletedTraining"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/me/trainings/{id}/certificate": {
      "get": {
        "tags": [
          "Trainings"
        ],
        "summary": "Eğitim sertifika indirme linki",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sertifika URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "certificate_not_issued"
          }
        }
      }
    },
...
  }
}